+++ packages/opentelemetry-instrumentation-langchain/tests/test_lcel.py
@@ -91,9 +91,34 @@ def test_simple_lcel(exporter):
     assert json.loads(
         prompt_task_span.attributes[SpanAttributes.TRACELOOP_ENTITY_OUTPUT]
     ) == {
-        "outputs": "messages=[SystemMessage(content='You are helpful assistant'), "
-        "HumanMessage(content='tell me a short joke')]",
         "kwargs": {"tags": ["seq:step:1", "test_tag"]},
+        "outputs": {
+            "id": ["langchain", "prompts", "chat", "ChatPromptValue"],
+            "kwargs": {
+                "messages": [
+                    {
+                        "id": ["langchain", "schema", "messages", "SystemMessage"],
+                        "kwargs": {
+                            "content": "You are helpful " "assistant",
+                            "type": "system",
+                        },
+                        "lc": 1,
+                        "type": "constructor",
+                    },
+                    {
+                        "id": ["langchain", "schema", "messages", "HumanMessage"],
+                        "kwargs": {
+                            "content": "tell me a short " "joke",
+                            "type": "human",
+                        },
+                        "lc": 1,
+                        "type": "constructor",
+                    },
+                ]
+            },
+            "lc": 1,
+            "type": "constructor",
+        },
     }