+++ packages/opentelemetry-instrumentation-openai/opentelemetry/instrumentation/openai/shared/chat_wrappers.py
@@ -1,3 +1,5 @@
+import asyncio
+import copy
 import json
 import logging
 import time
@@ -68,15 +70,15 @@ def chat_wrapper(
         SUPPRESS_LANGUAGE_MODEL_INSTRUMENTATION_KEY
     ):
         return wrapped(*args, **kwargs)
-
     # span needs to be opened and closed manually because the response is a generator
+
     span = tracer.start_span(
         SPAN_NAME,
         kind=SpanKind.CLIENT,
         attributes={SpanAttributes.LLM_REQUEST_TYPE: LLM_REQUEST_TYPE.value},
     )
 
-    _handle_request(span, kwargs, instance)
+    asyncio.run(_handle_request(span, kwargs, instance))
 
     try:
         start_time = time.time()
@@ -166,7 +168,7 @@ async def achat_wrapper(
         kind=SpanKind.CLIENT,
         attributes={SpanAttributes.LLM_REQUEST_TYPE: LLM_REQUEST_TYPE.value},
     )
-    _handle_request(span, kwargs, instance)
+    await _handle_request(span, kwargs, instance)
 
     try:
         start_time = time.time()
@@ -235,11 +237,11 @@ async def achat_wrapper(
 
 
 @dont_throw
-def _handle_request(span, kwargs, instance):
+async def _handle_request(span, kwargs, instance):
     _set_request_attributes(span, kwargs)
     _set_client_attributes(span, instance)
     if should_send_prompts():
-        _set_prompts(span, kwargs.get("messages"))
+        await _set_prompts(span, kwargs.get("messages"))
         if kwargs.get("functions"):
             _set_functions_attributes(span, kwargs.get("functions"))
         elif kwargs.get("tools"):
@@ -326,7 +328,32 @@ def _set_token_counter_metrics(token_counter, usage, shared_attributes):
             token_counter.record(val, attributes=attributes_with_token_type)
 
 
-def _set_prompts(span, messages):
+def _is_base64_image(item):
+    if not isinstance(item, dict):
+        return False
+
+    if not isinstance(item.get("image_url"), dict):
+        return False
+
+    if "data:image/" not in item.get("image_url", {}).get("url", ""):
+        return False
+
+    return True
+
+
+async def _process_image_item(item, trace_id, span_id, message_index, content_index):
+    if not Config.upload_base64_image:
+        return item
+
+    image_format = item["image_url"]["url"].split(";")[0].split("/")[1]
+    image_name = f"message_{message_index}_content_{content_index}.{image_format}"
+    base64_string = item["image_url"]["url"].split(",")[1]
+    url = await Config.upload_base64_image(trace_id, span_id, image_name, base64_string)
+
+    return {"type": "image_url", "image_url": {"url": url}}
+
+
+async def _set_prompts(span, messages):
     if not span.is_recording() or messages is None:
         return
 
@@ -335,8 +362,19 @@ def _set_prompts(span, messages):
 
         _set_span_attribute(span, f"{prefix}.role", msg.get("role"))
         if msg.get("content"):
-            content = msg.get("content")
+            content = copy.deepcopy(msg.get("content"))
             if isinstance(content, list):
+                content = [
+                    (
+                        await _process_image_item(
+                            item, span.context.trace_id, span.context.span_id, i, j
+                        )
+                        if _is_base64_image(item)
+                        else item
+                    )
+                    for j, item in enumerate(content)
+                ]
+
                 content = json.dumps(content)
             _set_span_attribute(span, f"{prefix}.content", content)
         if msg.get("tool_call_id"):