+++ packages/traceloop-sdk/traceloop/sdk/tracing/tracing.py
@@ -26,11 +26,12 @@ from opentelemetry.instrumentation.threading import ThreadingInstrumentor
 
 from opentelemetry.semconv_ai import SpanAttributes
 from traceloop.sdk import Telemetry
+from traceloop.sdk.images.image_uploader import ImageUploader
 from traceloop.sdk.instruments import Instruments
 from traceloop.sdk.tracing.content_allow_list import ContentAllowList
 from traceloop.sdk.utils import is_notebook
 from traceloop.sdk.utils.package_check import is_package_installed
-from typing import Dict, Optional, Set
+from typing import Callable, Dict, Optional, Set
 
 
 TRACER_NAME = "traceloop.tracer"
@@ -47,6 +48,7 @@ EXCLUDED_URLS = """
     posthog.com,
     sentry.io,
     bedrock-runtime,
+    sagemaker-runtime,
     googleapis.com,
     githubusercontent.com,
     openaipublic.blob.core.windows.net"""
@@ -58,6 +60,7 @@ class TracerWrapper(object):
     endpoint: str = None
     headers: Dict[str, str] = {}
     __tracer_provider: TracerProvider = None
+    __image_uploader: ImageUploader = None
 
     def __new__(
         cls,
@@ -67,12 +70,14 @@ class TracerWrapper(object):
         exporter: SpanExporter = None,
         should_enrich_metrics: bool = True,
         instruments: Optional[Set[Instruments]] = None,
+        image_uploader: ImageUploader = None,
     ) -> "TracerWrapper":
         if not hasattr(cls, "instance"):
             obj = cls.instance = super(TracerWrapper, cls).__new__(cls)
             if not TracerWrapper.endpoint:
                 return obj
 
+            obj.__image_uploader = image_uploader
             obj.__resource = Resource(attributes=TracerWrapper.resource_attributes)
             obj.__tracer_provider = init_tracer_provider(resource=obj.__resource)
             if processor:
@@ -125,18 +130,26 @@ class TracerWrapper(object):
 
             instrument_set = False
             if instruments is None:
-                init_instrumentations(should_enrich_metrics)
+                init_instrumentations(
+                    should_enrich_metrics, obj.__image_uploader.aupload_base64_image
+                )
                 instrument_set = True
             else:
                 for instrument in instruments:
                     if instrument == Instruments.OPENAI:
-                        if not init_openai_instrumentor(should_enrich_metrics):
+                        if not init_openai_instrumentor(
+                            should_enrich_metrics,
+                            obj.__image_uploader.aupload_base64_image,
+                        ):
                             print(Fore.RED + "Warning: OpenAI library does not exist.")
                             print(Fore.RESET)
                         else:
                             instrument_set = True
                     elif instrument == Instruments.ANTHROPIC:
-                        if not init_anthropic_instrumentor(should_enrich_metrics):
+                        if not init_anthropic_instrumentor(
+                            should_enrich_metrics,
+                            obj.__image_uploader.aupload_base64_image,
+                        ):
                             print(
                                 Fore.RED + "Warning: Anthropic library does not exist."
                             )
@@ -248,6 +261,12 @@ class TracerWrapper(object):
                     elif instrument == Instruments.BEDROCK:
                         if not init_bedrock_instrumentor(should_enrich_metrics):
                             print(Fore.RED + "Warning: Bedrock library does not exist.")
+                            print(Fore.RESET)
+                        else:
+                            instrument_set = True
+                    elif instrument == Instruments.SAGEMAKER:
+                        if not init_sagemaker_instrumentor(should_enrich_metrics):
+                            print(Fore.RED + "Warning: SageMaker library does not exist.")
                             print(Fore.RESET)
                         else:
                             instrument_set = True
@@ -525,9 +544,11 @@ def init_tracer_provider(resource: Resource) -> TracerProvider:
     return provider
 
 
-def init_instrumentations(should_enrich_metrics: bool):
-    init_openai_instrumentor(should_enrich_metrics)
-    init_anthropic_instrumentor(should_enrich_metrics)
+def init_instrumentations(
+    should_enrich_metrics: bool, base64_image_uploader: Callable[[str, str, str], str]
+):
+    init_openai_instrumentor(should_enrich_metrics, base64_image_uploader)
+    init_anthropic_instrumentor(should_enrich_metrics, base64_image_uploader)
     init_cohere_instrumentor()
     init_pinecone_instrumentor()
     init_qdrant_instrumentor()
@@ -546,6 +567,7 @@ def init_instrumentations(should_enrich_metrics: bool):
     init_urllib3_instrumentor()
     init_pymysql_instrumentor()
     init_bedrock_instrumentor(should_enrich_metrics)
+    init_sagemaker_instrumentor(should_enrich_metrics)
     init_replicate_instrumentor()
     init_vertexai_instrumentor()
     init_watsonx_instrumentor()
@@ -556,7 +578,9 @@ def init_instrumentations(should_enrich_metrics: bool):
     init_groq_instrumentor()
 
 
-def init_openai_instrumentor(should_enrich_metrics: bool):
+def init_openai_instrumentor(
+    should_enrich_metrics: bool, base64_image_uploader: Callable[[str, str, str], str]
+):
     try:
         if is_package_installed("openai"):
             Telemetry().capture("instrumentation:openai:init")
@@ -567,6 +591,7 @@ def init_openai_instrumentor(should_enrich_metrics: bool):
                 enrich_assistant=should_enrich_metrics,
                 enrich_token_usage=should_enrich_metrics,
                 get_common_metrics_attributes=metrics_common_attributes,
+                upload_base64_image=base64_image_uploader,
             )
             if not instrumentor.is_instrumented_by_opentelemetry:
                 instrumentor.instrument()
@@ -578,7 +603,9 @@ def init_openai_instrumentor(should_enrich_metrics: bool):
         return False
 
 
-def init_anthropic_instrumentor(should_enrich_metrics: bool):
+def init_anthropic_instrumentor(
+    should_enrich_metrics: bool, base64_image_uploader: Callable[[str, str, str], str]
+):
     try:
         if is_package_installed("anthropic"):
             Telemetry().capture("instrumentation:anthropic:init")
@@ -588,6 +615,7 @@ def init_anthropic_instrumentor(should_enrich_metrics: bool):
                 exception_logger=lambda e: Telemetry().log_exception(e),
                 enrich_token_usage=should_enrich_metrics,
                 get_common_metrics_attributes=metrics_common_attributes,
+                upload_base64_image=base64_image_uploader,
             )
             if not instrumentor.is_instrumented_by_opentelemetry:
                 instrumentor.instrument()
@@ -898,6 +926,24 @@ def init_bedrock_instrumentor(should_enrich_metrics: bool):
         return False
 
 
+def init_sagemaker_instrumentor(should_enrich_metrics: bool):
+    try:
+        if is_package_installed("boto3"):
+            from opentelemetry.instrumentation.sagemaker import SageMakerInstrumentor
+
+            instrumentor = SageMakerInstrumentor(
+                exception_logger=lambda e: Telemetry().log_exception(e),
+                enrich_token_usage=should_enrich_metrics,
+            )
+            if not instrumentor.is_instrumented_by_opentelemetry:
+                instrumentor.instrument()
+        return True
+    except Exception as e:
+        logging.error(f"Error initializing SageMaker instrumentor: {e}")
+        Telemetry().log_exception(e)
+        return False
+
+
 def init_replicate_instrumentor():
     try:
         if is_package_installed("replicate"):
@@ -918,7 +964,7 @@ def init_replicate_instrumentor():
 
 def init_vertexai_instrumentor():
     try:
-        if is_package_installed("vertexai"):
+        if is_package_installed("google-cloud-aiplatform"):
             Telemetry().capture("instrumentation:vertexai:init")
             from opentelemetry.instrumentation.vertexai import VertexAIInstrumentor
 
@@ -936,7 +982,9 @@ def init_vertexai_instrumentor():
 
 def init_watsonx_instrumentor():
     try:
-        if is_package_installed("ibm-watsonx-ai") or is_package_installed("ibm-watson-machine-learning"):
+        if is_package_installed("ibm-watsonx-ai") or is_package_installed(
+            "ibm-watson-machine-learning"
+        ):
             Telemetry().capture("instrumentation:watsonx:init")
             from opentelemetry.instrumentation.watsonx import WatsonxInstrumentor
 
@@ -1026,6 +1074,7 @@ def init_redis_instrumentor():
     try:
         if is_package_installed("redis"):
             from opentelemetry.instrumentation.redis import RedisInstrumentor
+
             instrumentor = RedisInstrumentor()
             if not instrumentor.is_instrumented_by_opentelemetry:
                 instrumentor.instrument(excluded_urls=EXCLUDED_URLS)